stationary trajectory with Steef's routines working
